A bunch of fixes for PrivSep, based on stuff found in the initial deployment on the CS pool:   - the old Starter now uses the Switchboard to create it's working     directory (it needs to because condor doesn't have the perms)   - fix a bunch of places where we weren't properly cleaning up     execute directories because we need to use the Switchboard for     this   - the StartD, in PrivSep mode, should not try to chmod the EXECUTE     directory since it doesn't have permission   - the Switchboard now conducts the recursive remove operation as     root so it can properly clean up   - the Switchboard now verifies the source UID on objects during the     recursive chown operation   - doubled the Switchboard's maximum config line to 8K; changing     Condor to not use large lines is a better solution, and should     be done in the near future   - the safe_dir_walk function in the Switchboard has been hardened     to relaibly not follow symlinks

+++ src/condor_starter.V5/user_proc.C
@@ -830,12 +830,26 @@ UserProc::delete_files()
 		do_unlink( core_name );
 	}
 
-	if( rmdir(local_dir) < 0 ) {
-		dprintf( D_ALWAYS,
-			"Can't remove directory \"%s\" - errno = %d\n", local_dir, errno
-		);
-	} else {
-		dprintf( D_ALWAYS, "Removed directory \"%s\"\n", local_dir );
+	dprintf( D_ALWAYS, "Removing directory \"%s\"\n", local_dir );
+	if (privsep_enabled()) {
+		// again, the PrivSep Switchboard expects a full path for
+		// the "remove dir" operation
+		//
+		MyString local_dir_path;
+		local_dir_path.sprintf("%s/%s", Execute, local_dir);
+		if (!privsep_remove_dir(local_dir_path.Value())) {
+			dprintf(D_ALWAYS,
+			        "privsep_remove_dir failed to remove %s\n",
+			        local_dir_path.Value());
+		}
+	}
+	else {
+		if( rmdir(local_dir) < 0 ) {
+			dprintf( D_ALWAYS,
+				"Can't remove directory \"%s\" - errno = %d\n",
+			        local_dir,
+			        errno);
+		}
 	}
 }
 
@@ -1437,8 +1451,21 @@ UserProc::UserProc( STARTUP_INFO &s ) :
 	buf.sprintf( "dir_%d", getpid() );
 	local_dir = new char [ buf.Length() + 1 ];
 	strcpy( local_dir, buf.Value() );
-	if( mkdir(local_dir,LOCAL_DIR_MODE) < 0 ) {
-		EXCEPT( "mkdir(%s,0%o)", local_dir, LOCAL_DIR_MODE );
+	if (privsep_enabled()) {
+		// the Switchboard expects a full path to privsep_create_dir
+		MyString local_dir_path;
+		local_dir_path.sprintf("%s/%s", Execute, local_dir);
+		if (!privsep_create_dir(get_condor_uid(), local_dir_path.Value())) {
+			EXCEPT("privsep_create_dir failure");
+		}
+		if (chmod(local_dir_path.Value(), LOCAL_DIR_MODE) == -1) {
+			EXCEPT("chmod failure after privsep_create_dir");
+		}
+	}
+	else {
+		if( mkdir(local_dir,LOCAL_DIR_MODE) < 0 ) {
+			EXCEPT( "mkdir(%s,0%o)", local_dir, LOCAL_DIR_MODE );
+		}
 	}
 	(void)umask(omask);
